unum-appstore/image-pipeline/thumbnail/requirements.txt
cfn-flip
boto3
Pillow
